Avoid integer overflow risk
When allocating chunks, we make them 10 times the preferred chunk size, if this is greater than the minimum chunk size.
We need to limit the size of the buffers that we pool, so that we prevent this 10x multiplication from overflowing int.
It also makes some sense to not pool extremely large buffers.
We use the same max chunk size as the old PooledBufferAllocator.
This gives us a cut-off point of ~100 MiB, where we don't pool buffers if they are bigger than this.