Increase errno cache and guard against IOOBE (#13254)
Motivation:

In the past we cached up to 512 errno values and used this cache to lookup the string representation without checking if the value if > 512. This value might be a bit low as more errno values are added over the time.

Modifications:

- Cache up to errno code 2048
- Use JNI call if errno code was not cached and so not throw an IOOBE.

Result:

No more exceptions on un-cached errno codes