Illustrate the right way to send a GET request (#13282)
Motivation:
The way to sent a GET request in the HttpUploadClient is wrong,  because if just send a GET request without a following LastHttpContent, the pipeline will be blocked, and  we can't reuse the channel to send another http request. The Example works because the channel is closed. but the following post and multi request both sent a LastHttpContent, so the get request should do this too. 
This may mislead the user.

Modification:
Fix the problem described in the motivation.

Result:
Illustrate the right way to send a GET request

---------

Co-authored-by: mipengcheng <mipengcheng3@jd.com>
Co-authored-by: Norman Maurer <norman_maurer@apple.com>