Export EPollEventLoop methods for closing/opening native file descriptors as UnstableAPI
Motivation:

This change intends to support an application using Netty with native epoll (e.g. Quarkus app) to perform the Checkpoint and Restore on JVM implementing this, specifically using OpenJDK CRaC or future versions of OpenJDK.

Modification:

The biggest risk factor here is that formerly final fields with file descriptors can change; theoretically this could affect performance in absence of checkpoint/restore process. However most likely a modern JVM is smart enough to assume that the field does not change, and optimistically consider it constant anyway.

Result:

There is no change in functionality.

Signed-off-by: Radim Vansa <rvansa@azul.com>