A few more fix-ups after the logging changes in #13285
Motivation:
The is _a lot_ going on in #13285, and we missed that we exposed the SLF4J level in a couple of places.
We also _somehow_ missed that we broke compilation of EchoIT in the buffer-memory-segment module.

Modification:
Replace the remaining places that exposed SLF4J Level to the public API, and fix the compilation.

Result:
No more left-overs to clean up after #13285