Reset byte buffer in loop for AbstractDiskHttpData.setContent.
Motivation:
When the number of bytes available from a single InputStream.read-call
varies between invocations (like can be the case for resources served from within a jar),
the first invocation of flip() sets the limit to this smaller value.
As InputStream.read(byte[]) is used, that call does not see the new limit and might read more bytes
causing the next call to ByteBuffer.position() to fail. Resetting the limit in between resolves this issue.