Add support for local and remote addresses on the server for child channels when UDS (#13323)
Motivation:
Local and remote addresses are not available on the server for child channels when
Unix Domain Sockets is used. There are use cases where one may need this information
in order to collect metrics per local or remote addresses.

Modifications:
- Add JNI code for obtaining local/remote address
- When creating a child channel initialise the local/remote addresses.
- Add junit tests

Result:
Local and remote addresses are now available on the server for child channels when
Unix Domain Sockets is used.

Co-Authored-by: Norman Maurer <norman_maurer@apple.com>