Do not throw exception when resolveCache and TTLs are both configured
Motivation:
The configured DnsNameResolverBuilder#TTLs (minTtl, maxTtl and negativeTtl) are used for creating
default resolveCache, cnameCache and authoritativeDnsServerCache. Configuring a custom
resolveCache(or cnameCache, or authoritativeDnsServerCache) does not mean that TTLs are not needed anymore
for the rest of the caches that are not configured.
With the current implementation there is no way for the user to specify a custom resolveCache,
but to leave Netty to create cnameCache and authoritativeDnsServerCache.

Modifications:
- DnsNameResolverBuilder does not throw IllegalStateException, but logs when a custom
resolveCache/cnameCache/authoritativeDnsServerCache and TTLs are configured, indicating
that for the custom resolveCache/cnameCache/authoritativeDnsServerCache, the TTLs will be ignored.
- junit tests are added

Result:
Fixes #13384