HTTP/2 don't propagate `ChannelInputShutdownEvent` to active streams
Motivation:

Follow-up for #13394 to remove `ChannelInputShutdownEvent` propagation.
Discussed with @Scottmitch, `ChannelInputShutdownEvent` is a premature
event, most users should rely on `ChannelInputShutdownReadComplete`
instead. Let's propagate only the later one and if there is a use-case
for `ChannelInputShutdownEvent` we can add its propagation later.

Modifications:

- Remove `CHANNEL_INPUT_SHUTDOWN_EVENT_VISITOR` to stop propagation of
`ChannelInputShutdownEvent`;
- Adjust tests;

Result:

`ChannelInputShutdownEvent` is not propagated for now. We can add it
back when there is a use-case.