Consolidate flushes if possible (#518)
Motivation:

While we are in the channelReadComplete(...) method we might trigger
flush() multiple times. This is wasteful as we will most likely do
multiple sendto syscalls. We can do a better job by consolidate these
flushes to just one and so be able to replace X sendto syscalls with 1
sendmmsg syscall.

Modifications:

- Keep track of if we are currently in the channelReadComplete(...)
method and if so try to consolidate flushes.

Result:

Less syscalls