Always enable leak tracking for derived buffers if parent is tracked
Motivation:

Currently, when leak tracking is set to SIMPLE or ADVANCED, and a
derived buffer is created, the buffer is selected for leak tracking if
the parent buffer is tracked and the derived buffer is also selected by
an independent roll of the dice. As a result, deeply derived buffers are
exponentially less likely to be selected for leak tracking. This makes
it difficult to find and diagnose leaks related to such buffers.

Modifications:

When a derived buffer is created, and the parent buffer has leak
tracking, the derived buffer now always has leak tracking applied.

Result:

Leaks of derived buffers are easier to find.