Implement contains for SelectedSelectionKeySet (#13452)
Motivation:
The `contains` method may be called in the
`sub.nio.ch.SelectorImpl.processReadyEvents` method on macOS and
Windows, so we need to implement it correctly. Otherwise, we can lose
read-poll events.
See the investigation on the Hazelcast repo:
https://github.com/hazelcast/hazelcast/pull/24267

Modification:
Add a contains method that just does a linear scan for the selection
key. This preserves the performance characteristics of all the other
methods.

Result:
Fixes #13328