Give explicit error on missing BouncyCastle PKIX library (#13475)
Motivation:
If the BounyCastle PKIX library is missing from the classpath, Netty
throws an exception like this: `Failed to generate a self-signed X.509
certificate using Bouncy Castle`. This is not really helpful as the user
has to go through Stacktrace to find out what went wrong. To combat
this, I added a util method that checks if
`org.bouncycastle.cert.X509v3CertificateBuilder` is available in the
classpath or not. Based on this, we can return explicit clear exceptions
which helps in debugging.

`org.bouncycastle.cert.X509v3CertificateBuilder` is part of the
BounyCastle PKIX library.

Modification:
Added `isBouncyCastleAvailable()` method to check for
`org.bouncycastle.cert.X509v3CertificateBuilder` availability in
classpath.

Result:
Clear explicit error on missing BouncyCastle PKIX library