Add Java 21 support and enable Panama build
Motivation:
The panama snapshot build was disabled because of a compilation error in upstream panama.
Let's try to bring it back and see if it works.

Modification:
Update the project `pom.xml` file to support Java 21.
Update the panama workflow to use JDK 20 as build-jdk (used for building Panama), and enable the panama build workflow again.

Result:
Maybe this build works again.