Separate encode methods for CookiePair and SetCookie Motivation:
Before this patch, the `HttpCookiePair.encoded()` would return the `Cookie` syntax for `HttpCookiePair`, but the `Set-Cookie` syntax for `HttpSetCookie`. `DefaultHttpHeaders.addCookie` trusts this value, so passing a `HttpSetCookie` to the method would produce and invalid `Cookie` header.

There is really no need for these two formats to use the same method.

Modification:

Replace the `encoded` method with separate `encodedCookie` and `encodedSetCookie` methods for the respective header formats.

Result:

`HttpSetCookie` instances can now be used with `addCookie` without issue.