Enable validation for HttpHeaders.emptyHeaders() Motivation:
While `emptyHeaders()` should usually stay empty, this is not enforced in netty 5. Thus, it is preferable to enable validation by default, in case the header map *is* modified.

Modification:

Enable validation for `emptyHeaders()`.

Result:

If the `emptyHeaders()` instance is wrongly modified, the values will be validated properly. This should have no performance impact, as validation is only done on modification, which ideally doesn't happen for `emptyHeaders()`.