Handle wildcard addresses in PcapWriteHandler Motivation:
If a UDP socket is bound to :: or 0.0.0.0, on some systems, it can still send packets to addresses of the other address family. This leads to a mismatch between the local and remote address families, that PcapWriteHandler cannot handle.

Modification:

In this scenario, coerce the local wildcard address to the wildcard address of the proper family.

Result:

Local wildcard address is handled properly and packets are logged as expected.