Changes lazily-initialized 'codeClass' to final field to simplify thread safety of HttpResponseStatus.
Motivation:

Simplify thread safety of HttpResponseStatus by avoiding multiple threads
entering "check if null" branch in `codeClass()` and fix thread sanitizer
warnings. There is a small performance penalty from always computing the
`codeClass` in the constructor, but in practice `codeClass()` is essentially
always used in the codecs so we are just deferring the computation cost until
later, and final fields are much easier to reason about.

Modifications:

Changes lazily-initialized 'codeClass' to final field initialized in
constructor and changes getter to return final field value.

Result:

Thread safety of HttpResponseStatus is simpler and thread sanitizer checks are
happy (this is how we found this in the first place).