Revert "Make initializer as `static` (#13506)"
Motivations:

By changing the code to use a static block we did remove the ability to load the class at all. This is sort of a regression as it may break existing code.

Modifications:

This reverts commit 1c7f0faac060313bfd313d0a50fdf83a2fd2c9c9.

Result:

It's possible again to at least load the class all the time as before.
Fixes https://github.com/netty/netty/issues/13523