SSLSession.getPeerCertificateChain() should throw UnsupportedOperationException if javax.security.cert.X509Certificate can bot be created
Motivation:

SSLSession.getPeerCertificateChain() is allowed to throw UnsupportedOperationException if the implementation does not support it. We should do exactly this if the JDK that is used can not create a cert from a byte array. Otherwise we will produce an exception much later which is confusing.

Modifications:

Check if we can create such a cert and so support it, if not throw an UnuspportedOperationException as soon as the user tries to use the method.

Result:

Fixes https://github.com/netty/netty/issues/13560