Add service-loaded extension points for channel initialization
**Motivation:**

Netty is used by many libraries and frameworks that (reasonably) hide the fact that they use Netty under the covers.
When Netty is hidden, it is hard, or sometimes impossible, to modify the channel pipelines, attributes, or options, from the outside.
It might not even be clear if a framework or library makes use of Netty at all.
However, we sometimes want to apply some changes, or make some checks, to most or all channels that are initialized by Netty, regardless of the framework or library that is using Netty in the given case.

Some examples of use-cases are:
- Web application firewalls.
- Server-side request forgery filters.
- Intrusion detection.
- Metrics gathering.

To address these use-cases in a way that don't require integrators to somehow find every Netty usage in a process, we introduce a service-loaded extension point that hooks into the channel initialization process.

**Modification:**

A new interface, `ChannelInitializerExtension`, is added.
By default, implementations of this interface are found and service-loaded at runtime by `AbstractBootstrap`, with the help of hte `ChannelInitializerExtensions` utility class.

The interface offers three callback methods, one for the initialization of each of the different "kinds" of channels:
- Server listener channels.
- Server child channels.
- Client channels.

The extension also gets the opportunity to decide if it is applicable to a particular context or use-case, through the `isApplicable` method.
The method takes an `ApplicableInfo` object, which is created by `AbstractBootstrap`.
The `ApplicableInfo` is a parameter object, and a `final` class, which allow us to add information to it in the future by adding more methods to it, without breaking backwards compatibility.
To begin with, we only expose the concrete bootstrap class through the `ApplicableInfo`, so extensions can tell if they are being considered for a server, or a client usage.

By default, we service-load all extensions that are available at runtime.
However, integrators can disable the use of extensions on a per-bootstrap basis with the new `disableChannelInitializerExtensions` method.

Extensions can also be disabled for the entire JVM process by setting the `io.netty.bootstrap.extensions` system property to `empty` or `false`.
This system property is intentionally defined as a string, so we can offer more controls in the future.
For instance, we may wish to offer injecting different loader mechanisms, in which case the property could denote a class name.

**Result:**

We have extension points for use cases that are cutting-across different Netty uses, and require low-level Netty access.
This makes it relatively easy to implement such use cases without requiring deep access to the Netty internals of arbitrary libraries and frameworks, and without even needing to know which libraries and frameworks are actually using Netty.
The only restriction to this, is the use of shading with class-relocation, which in each case would require their own `META-INF/services` file.
Thankfully, most libraries and frameworks that shade Netty, also offer non-shading versions.