Only coalesce when there is more than one inbound buffer
Motivation:
In #13568 we added packet coalescing to the local transport.
This will try to coalesce every time we observe a Buffer in the inboundBuffer queue.
We should, however, only try to coalesce when there is at least two buffers in a row in the inbound buffer queue.
Otherwise, we will end up doing an allocation and memory copy without actually reducing the number of buffers sent down the pipeline.

Modification:
Only go down the coalescing code path if we not only poll a buffer from the queue, but peek tells us that we have another buffer waiting after the first.

Result:
Avoid useless coalescing work when we only have one buffer in the inbound buffer queue.