Make CipherSuiteConverter also cache failures (#13591)
Motivation:

When instantiation lots of SslContexts (load test simulating lots of
distinct user agents), CipherSuiteConverter performs lots of extra work
(regex) against all the Java cipher suites that can't be converted to
OpenSSL (BoringSSL in my case).

The reason is CipherSuiteConverter only caches successes and not
failures.

Modifications:

* Introduce `CachedValue` that lets store nullable values in a CHM
* always cache the result of `toOpenSslUncached`

Result:

Values that are known to not be convertible to OpenSSL are cached
instead of being tried over and over again.