Lazy tableSizeChangeOutput (#13589)
Motivation:

`tableSizeChangeOutput` is an unpooled ByteBuf which is eagerly
allocated for every `DefaultHttp2HeadersEncoder`.

We only need it when we get a SETTINGS frame that has this header, which
might never be the case depending on the application.

This result in excessive allocations when having to deal with short
lived HTTP/2 connections under load.

Modification:

Only allocate `tableSizeChangeOutput` if needed.

Result:

Less allocations