Add reset API to IdleStateHandler Motivation:
Sometimes it is useful to reset the read/write timeout of an IdleStateHandler without an actual read/write event. At the moment, this requires a complicated dance of sending a manual read/write event and then intercepting it again to avoid triggering downstream handlers.

Modification:

Add API methods to explicitly reset the timeouts. Add tests for the new API. Fix the `anyNotIdle` test utility method, because it turns out it did not actually confirm the action reset the timeout.

Result:

The timeouts can be reset manually by the developer without resorting to workarounds.