Revert "HttpHeaderValidationUtil should reject chars past the 1 byte range" (#13615)
Motivation:
[RFC 9110 Section
5.5-4](https://datatracker.ietf.org/doc/html/rfc9110#section-5.5-4) says
we _should_ constrain ourselves to only putting US-ASCII in header
values, _but_ encodings beyond that are technically allowed as long as
they are ASCII-compatible (e.g. ISO-8859-1 or UTF-8).
In such a case, some of their characters will fall into the "obs-text"
clause.

One complication with our header validation is that we don't at that
point know if we're the recipient or the sender.
And we also don't at that point know what encoding is used.

The Netty header encoder clobbers non-ASCII characters, but people may
have custom header encoders that permit more characters.

Modification:
This reverts #13541, so we no longer enforce that `String` header values
only contain US-ASCII characters.

Result:
Restore the "support" for non-ASCII header encodings that we had
previously.