Check buffer length before reading DTLS record length
Adds a check to make sure the buffer has the required bytes to read
before we try and get the DTLS record's length. This also adds a test to
make sure we will return NOT_ENOUGH_DATA in the event we don't have
enough bytes to get to the DTLS record's length.