Added check for pseudo-headers in trailers (#13603)
Motivation:
According to the RFC for HTTP2 on Section 8.1.2.1, 'Pseudo-header fields
MUST NOT appear in trailers. Endpoints MUST treat a request or response
that contains undefined or invalid pseudo-header fields as malformed
(Section 8.1.2.6).'

Modifications:
Updated Http2Decoder to check trailers for Pseudo-Headers and treat them
as invalid requests according to RFC 7540 Section 8.1.2.6

Result:
After this change, requests with Pseudo-Headers sent inside trailers
will be treated as malformed and a stream error is thrown.

Fixes: #13602

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>
Co-authored-by: Bryce Anderson <bryce_anderson@apple.com>