Support DTLS headers when parsing encrypted packet lengths #2 (#13656)
Motivation:

Supporting DTLS headers allows users to use a DTLS SSLContext to build a
SSLEngine for a SslHandler. There are probably more changes that need to
be added to handle edge cases in DTLS, but this appears to work and
starts the process of supporting DTLS.

Modifications:

Checks for the DTLS version after failing to find the TLS version. If
DTLS version is found, we read the length of each DTLS record until
there are no more records.

Result:

Users can create a SslHandler with a DTLS SSLContext and connect to a
simple server using DTLS.

---------

Co-authored-by: David Stocking <david.stocking@stryker.com>
Co-authored-by: Norman Maurer <norman_maurer@apple.com>