Add service-loaded extension points for channel initialization (#13565)
**Motivation:**

Netty is used by many libraries and frameworks that (reasonably) hide
the fact that they use Netty under the covers. When Netty is hidden, it
is hard, or sometimes impossible, to modify the channel pipelines,
attributes, or options, from the outside. It might not even be clear if
a framework or library makes use of Netty at all. However, we sometimes
want to apply some changes, or make some checks, to most or all channels
that are initialized by Netty, regardless of the framework or library
that is using Netty in the given case.

Some examples of use-cases are:
- Web application firewalls.
- Server-side request forgery filters.
- Intrusion detection.
- Metrics gathering.

To address these use-cases in a way that don't require integrators to
somehow find every Netty usage in a process, we introduce a
service-loaded extension point that hooks into the channel
initialization process.

**Modification:**

A new abstract class, `ChannelInitializerExtension`, is added.
Implementations of this interface can be found and service-loaded at
runtime by `AbstractBootstrap`, with the help of the
`ChannelInitializerExtensions` utility class.

The `ChannelInitializerExtension` class offers three callback methods,
one for the initialization of each of the different "kinds" of channels:
- Server listener channels.
- Server child channels.
- Client channels.

The extensions are disabled by default for security reasons, and require
opt in by setting the `io.netty5.bootstrap.extensions` system property to
`serviceload`.
To see what extensions are available, the system property can be set to
`log`, and we will load and log (at INFO level) what extensions are
available, but not actually run them.

**Result:**

We have extension points for use cases that are cutting-across different
Netty uses, and require low-level Netty access. This makes it relatively
easy to implement such use cases without requiring deep access to the
Netty internals of arbitrary libraries and frameworks, and without even
needing to know which libraries and frameworks are actually using Netty.
The only restriction to this, is the use of shading with
class-relocation, which in each case would require their own
`META-INF/services` file. Thankfully, most libraries and frameworks that
shade Netty, also offer non-shading versions.