Fix some bugs in the IOUringHandler
Motivation:

It's possible for the event FD to be released an
yet have a new submission write to the FD, which
may be reallocated by the OS in the meantime.

Modifications:

- Properly shutdown the wakeup pathway and drain
  any pending/future writes to the event fd.

Result:

An IOUring event loop will no longer be able to
(wrongly) write an event to a fd which has already
been closed, and potentially reallocated.