Add check for JAVA8_HOME to release script
Motivation:
This environment variable is required.
If it's not set, the Netty release will end up being built wrong bytecode version and incompatible ABI.

Modification:
Add a check that the JAVA8_HOME environment variable does not resolve to the empty string.

Result:
We're less likely to get a broken release from mistakes in the manual steps.